--- src/components/ui/stories/alert_dialog_demo.stories.tsx
@@ -1,50 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react'
-
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from '../alert-dialog'
-
-const meta = {
-  title: 'Shadcn/alert',
-  component: () => {
-    return (
-      <>
-        <AlertDialog>
-          <AlertDialogTrigger>Open</AlertDialogTrigger>
-          <AlertDialogContent>
-            <AlertDialogHeader>
-              <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
-              <AlertDialogDescription>
-                This action cannot be undone. This will permanently delete your
-                account and remove your data from our servers.
-              </AlertDialogDescription>
-            </AlertDialogHeader>
-            <AlertDialogFooter>
-              <AlertDialogCancel>Cancel</AlertDialogCancel>
-              <AlertDialogAction>Continue</AlertDialogAction>
-            </AlertDialogFooter>
-          </AlertDialogContent>
-        </AlertDialog>
-      </>
-    )
-  },
-  parameters: {
-    layout: 'centered',
-  },
-} satisfies Meta<{}>
-
-export default meta
-
-type Story = StoryObj<{}>
-
-export const AlertDialogDemo: Story = {
-  args: {},
-}
